==> Analyses/01_processing_of_data_entries.Rmd <==
---
title: "TestCode"
author: "Vera Weisbecker"
date: "8 December 2018"
output: html_document
---



#The below summarizes the table according to the values given:

#For missing average male and female values:average the min and max values and enter into m/f average values; 

#For missing min/max values: if available, use the maximal values within both males and females for a species min/max;

#For missing overall average values: if there is an average value for males and females (regardles of given or derived from within-sex min/maxes), use their averages to enter into the final values; if there is none, use the average of min/max 

#if there are min max values for either sex, enter their overall min and max into the min/max columns.

#Note that this chunk returns an error if run on an empty environment (but not upon running again) in Rstudio but it does not affect the output; it's some kind of internal name conflict but the output is 100% correct

```{r }


data=read.csv("../Data/raw_data.csv", header=TRUE)

#make a matrix for all data but also add 3 columns for final Tl, Bl and Wt
#data_processed<-matrix(ncol=(ncol(data)), nrow=nrow(data))
#colnames(data_processed)<-c(colnames(data))
#rownames(data_processed)<-data$Species

# turn into data frame to feed into loop; 
data_processed<-data

#if there are male/female ranges, provide their mean; this is clunky but I prefer to have the names written out to reduce human error in assigning column numbers
for (i in 1:length(data$Species)){
  
    
    #if there are min/max values for males and females, average them unless an average is already given; this code here separates M/F because sometimes ranges are only given for one sex
    if(is.na(data$Wt_min_F[i])==FALSE & is.na(data$Wt_Av_F[i]==TRUE)){
  
      data_processed$Wt_Av_F[i]<-mean(c(data$Wt_min_F[i],data$Wt_max_F[i]))
    }
    
    if(is.na(data$Wt_min_M[i])==FALSE & is.na(data$Wt_Av_M[i]==TRUE)){
  
      data_processed$Wt_Av_M[i]<-mean(c(data$Wt_min_M[i],data$Wt_max_M[i]))
      
    }
  
    if(is.na(data$Bl_min_F[i])==FALSE & is.na(data$Bl_Av_F[i]==TRUE)){
  
      data_processed$Bl_Av_F[i]<-mean(c(data$Bl_min_F[i],data$Bl_max_F[i]))
    }
    
    if(is.na(data$Bl_min_M[i])==FALSE & is.na(data$Bl_Av_M[i]==TRUE)){
  
      data_processed$Bl_Av_M[i]<-mean(c(data$Bl_min_M[i],data$Bl_max_M[i]))
      
    }
  
    if(is.na(data$Tl_min_F[i])==FALSE & is.na(data$Tl_Av_F[i]==TRUE)){
  
      data_processed$Tl_Av_F[i]<-mean(c(data$Tl_min_F[i],data$Tl_max_F[i]))
    }
    
    if(is.na(data$Tl_min_M[i])==FALSE & is.na(data$Tl_Av_M[i]==TRUE)){
  
      data_processed$Tl_Av_M[i]<-mean(c(data$Tl_min_M[i],data$Tl_max_M[i]))
      
    }
 
  # if males or females have ranges, enter the max and min for each species in the relevant column; this assumes that overall ranges are never given together with the more informative range values by sex 

    if((is.na(data$Wt_min_F[i])==FALSE | is.na(data$Wt_min_M[i])==FALSE)){
  
      data_processed$Wt_min[i]<-min(c(data$Wt_min_F[i],data$Wt_max_F[i],data$Wt_min_M[i],data$Wt_max_M[i]))
      data_processed$Wt_max[i]<-max(c(data$Wt_max_F[i],data$Wt_max_F[i],data$Wt_max_M[i],data$Wt_max_M[i]))
    }
  
    if(is.na(data$Bl_min_F[i])==FALSE | is.na(data$Bl_min_M[i])==FALSE){
  
      data_processed$Bl_min[i]<-min(c(data$Bl_min_F[i],data$Bl_max_F[i],data$Bl_min_M[i],data$Bl_max_M[i]))
      data_processed$Bl_max[i]<-max(c(data$Bl_max_F[i],data$Bl_max_F[i],data$Bl_max_M[i],data$Bl_max_M[i]))
    }
  
    if(is.na(data$Tl_min_F[i])==FALSE | is.na(data$Tl_min_M[i])==FALSE){
  
      data_processed$Tl_min[i]<-min(c(data$Tl_min_F[i],data$Tl_max_F[i],data$Tl_min_M[i],data$Tl_max_M[i]))
      data_processed$Tl_max[i]<-max(c(data$Tl_max_F[i],data$Tl_max_F[i],data$Tl_max_M[i],data$Tl_max_M[i]))
    }
    
}

#now re-process data_processed to fill NAs in the species means, first by averaging male/female values, second (if sex averages are not available) by averaging sex-less min/max values. This overwrites the first pass summary version of data_processed with the second pass summary version

for (i in 1:(nrow(data_processed))){
  
   
      #if there is a male/female average value, use its average as final value (assuming that female ranges also mean that there is a male range)
      if(is.na(data_processed$Wt_g[i])==TRUE & is.na(data_processed$Wt_Av_F[i])==FALSE){
  
      data_processed$Wt_g[i]<-mean(c(data_processed$Wt_Av_F[i], data_processed$Wt_Av_M[i]))
      }
  
      if(is.na(data_processed$Bl_mm[i])==TRUE & is.na(data_processed$Bl_Av_F[i])==FALSE){
  
      data_processed$Bl_mm[i]<-mean(c(data_processed$Bl_Av_F[i], data_processed$Bl_Av_M[i]))
      }
    
      if(is.na(data_processed$Tl_mm[i])==TRUE & is.na(data_processed$Tl_Av_F[i])==FALSE){
  
      data_processed$Tl_mm[i]<-mean(c(data_processed$Tl_Av_F[i], data_processed$Tl_Av_M[i]))
      }
  
  
      #if there are no male/female average values (regardless of newly created above or originally in the book; i.e. the male and female), average the ranges
      if(is.na(data_processed$Wt_g[i])==TRUE & is.na(data_processed$Wt_Av_F[i])==TRUE){
  
      data_processed$Wt_g[i]<-mean(c(data_processed$Wt_max[i], data_processed$Wt_min[i]))
      }
  
      if(is.na(data_processed$Bl_mm[i])==TRUE & is.na(data_processed$Bl_Av_F[i])==TRUE){
  
      data_processed$Bl_mm[i]<-mean(c(data_processed$Bl_max[i], data_processed$Bl_min[i]))
      }
    
      if(is.na(data_processed$Tl_mm[i])==TRUE & is.na(data_processed$Tl_Av_F[i])==TRUE){
  
      data_processed$Tl_mm[i]<-mean(c(data_processed$Tl_max[i], data_processed$Tl_min[i]))
      }  
}

#clean up environment
remove(i)

#Write the csv; row.names need to be FALSE because otherwise it saves with an extra column of numbers
write.csv(data_processed, file="../Data/Full_dataset.csv", row.names=FALSE)

```

#A second, smaller dataset that just computes the full table for species where male and female averages and ranges are known. This is for double-checking that the midpoints and means are comparable

##Separate species from raw data that have m/f ranges and means;THEN compute their mean and overall min/max; THEN compute the mean of the overall min max and compare to species mean

```{r}

data_for_mid_checks<-data

#Species that have both range and mean data for males/females or overall for at least one of the values

data_for_mid_checks <- data_for_mid_checks  [which(
      
      #Males/females
      is.na(data$Wt_min_F)==FALSE & is.na(data$Wt_Av_F)==FALSE | 
      is.na(data$Wt_min_M)==FALSE & is.na(data$Wt_Av_M)==FALSE |
      is.na(data$Bl_min_F)==FALSE & is.na(data$Bl_Av_F)==FALSE | 
      is.na(data$Bl_min_M)==FALSE & is.na(data$Bl_Av_M)==FALSE |
      is.na(data$Tl_min_F)==FALSE & is.na(data$Tl_Av_F)==FALSE | 
      is.na(data$Tl_min_M)==FALSE & is.na(data$Tl_Av_M)==FALSE |
      
      #Overall
      is.na(data$Wt_min)==FALSE & is.na(data$Wt_g)==FALSE      |
      is.na(data$Bl_min)==FALSE & is.na(data$Bl_mm)==FALSE     | 
      is.na(data$Tl_min)==FALSE & is.na(data$Tl_mm)==FALSE  
      ),

]

#Creating whole-species overall ranges - so the absolute min and max for males and females combined

mean_range_data <-data_for_mid_checks

for (i in 1:length(data_for_mid_checks$Species)){
  
    if((is.na(data_for_mid_checks$Wt_min_F[i])==FALSE | is.na(data_for_mid_checks$Wt_min_M[i])==FALSE)){
  
      mean_range_data$Wt_min[i]<-min(c(data_for_mid_checks$Wt_min_F[i],data_for_mid_checks$Wt_max_F[i],data_for_mid_checks$Wt_min_M[i],data_for_mid_checks$Wt_max_M[i]))
      mean_range_data$Wt_max[i]<-max(c(data_for_mid_checks$Wt_max_F[i],data_for_mid_checks$Wt_max_F[i],data_for_mid_checks$Wt_max_M[i],data_for_mid_checks$Wt_max_M[i]))
    }
  
    if(is.na(data_for_mid_checks$Bl_min_F[i])==FALSE | is.na(data_for_mid_checks$Bl_min_M[i])==FALSE){
  
      mean_range_data$Bl_min[i]<-min(c(data_for_mid_checks$Bl_min_F[i],data_for_mid_checks$Bl_max_F[i],data_for_mid_checks$Bl_min_M[i],data_for_mid_checks$Bl_max_M[i]))
      mean_range_data$Bl_max[i]<-max(c(data_for_mid_checks$Bl_max_F[i],data_for_mid_checks$Bl_max_F[i],data_for_mid_checks$Bl_max_M[i],data_for_mid_checks$Bl_max_M[i]))
    }
  
    if(is.na(data_for_mid_checks$Tl_min_F[i])==FALSE | is.na(data_for_mid_checks$Tl_min_M[i])==FALSE){
  
      mean_range_data$Tl_min[i]<-min(c(data_for_mid_checks$Tl_min_F[i],data_for_mid_checks$Tl_max_F[i],data_for_mid_checks$Tl_min_M[i],data_for_mid_checks$Tl_max_M[i]))
      mean_range_data$Tl_max[i]<-max(c(data_for_mid_checks$Tl_max_F[i],data_for_mid_checks$Tl_max_F[i],data_for_mid_checks$Tl_max_M[i],data_for_mid_checks$Tl_max_M[i]))
    }
    
}


#now averaging male/female average values and entering as species averages; so now we have all data that have a confident average.

for (i in 1:(nrow(mean_range_data))){
  
   
      #if there is a male/female average value, use its average as final value (assuming that female ranges also mean that there is a male range)
      if(is.na(mean_range_data$Wt_g[i])==TRUE & is.na(mean_range_data$Wt_Av_F[i])==FALSE){
  
      mean_range_data$Wt_g[i]<-mean(c(mean_range_data$Wt_Av_F[i], mean_range_data$Wt_Av_M[i]))
      }
  
      if(is.na(mean_range_data$Bl_mm[i])==TRUE & is.na(mean_range_data$Bl_Av_F[i])==FALSE){
  
      mean_range_data$Bl_mm[i]<-mean(c(mean_range_data$Bl_Av_F[i], mean_range_data$Bl_Av_M[i]))
      }
    
      if(is.na(mean_range_data$Tl_mm[i])==TRUE & is.na(mean_range_data$Tl_Av_F[i])==FALSE){
  
      mean_range_data$Tl_mm[i]<-mean(c(mean_range_data$Tl_Av_F[i], mean_range_data$Tl_Av_M[i]))
      }
}
 
#Now create and fill columns with overall range averages; but where m/f ranges area available, these will be overwritten in the next step as these are better data and available in most cases

mean_range_data$range_mid_Bl <-  rowMeans(mean_range_data[c('Bl_min', 'Bl_max')], na.rm=TRUE)
mean_range_data$range_mid_Tl <-  rowMeans(mean_range_data[c('Tl_min', 'Tl_max')], na.rm=TRUE)
mean_range_data$range_mid_Wt <-  rowMeans(mean_range_data[c('Wt_min', 'Wt_max')], na.rm=TRUE)


# where ranges between males and females are available, averaging all four to create a midpoint


for (i in 1:length(mean_range_data$Species)){
  
    if((is.na(mean_range_data$Wt_min_F[i])==FALSE | is.na(mean_range_data$Wt_min_M[i])==FALSE)){
  
      mean_range_data$range_mid_Wt[i]<-min(c(mean_range_data$Wt_min_F[i],mean_range_data$Wt_max_F[i],mean_range_data$Wt_min_M[i],mean_range_data$Wt_max_M[i]))
      
    }
  
    if(is.na(mean_range_data$Bl_min_F[i])==FALSE | is.na(mean_range_data$Bl_min_M[i])==FALSE){
  
      mean_range_data$range_mid_Bl[i]<-min(c(mean_range_data$Bl_min_F[i],mean_range_data$Bl_max_F[i],mean_range_data$Bl_min_M[i],mean_range_data$Bl_max_M[i]))
      
    }
  
    if(is.na(mean_range_data$Tl_min_F[i])==FALSE | is.na(mean_range_data$Tl_min_M[i])==FALSE){
  
      mean_range_data$range_mid_Tl[i]<-min(c(mean_range_data$Tl_min_F[i],mean_range_data$Tl_max_F[i],mean_range_data$Tl_min_M[i],mean_range_data$Tl_max_M[i]))
      
    }
    
}

  
#clean up environment
remove(i)

#Write the csv; row.names need to be FALSE because otherwise it saves with an extra column of numbers
write.csv(mean_range_data, file="../Data/means_test_dataset.csv", row.names=FALSE)

```
