==> Analyses/02_preparation_for_analyses.Rmd <==
---
title: "Tail study"
author: "Anonymous - available on Github"
date: "27 February 2019"
output: html_document
---


## Load required packages and data
```{r, warnings=FALSE}
library(ape)
library(geiger)

```

##prepare data and trees; this includes deleting entries with NAS in the main averages
```{r}
# The data, partition between marsupials and rodents
data=read.csv("../data/Full_dataset.csv")

#log10 transform all numerical data

#extract numerical columns
numcols=as.data.frame(data[,2:28])

#turn NAs into 1 so they turn into easily replaceable zeroes without messing with the logging
numcols[is.na(numcols)]<-1

#log transform
numcols_log10=log10(numcols)

#replace 0s with NAS again
numcols_log10[numcols_log10==0]<-NA

#replace original data with logged data. +1 because 2:34 is 33. There is probably a more elegant way of doing this

data[,2:(ncol(numcols)+1)]=numcols_log10

#partition datasets; this cannot be automated easily so need to check for accuracy by looking at dataset

#exclude wombats,koalas and marsupial moles who are outliers using grep (which returns the number of the row that has the species name).

data=data[-(c(grep("Vombatus_ursinus", data$Species),grep("Phascolarctos_cinereus", data$Species),grep("Lasiorhinus_latifrons", data$Species), grep("Notoryctes_typhlops", data$Species), grep("Notoryctes_caurinus", data$Species))),]

data$Clade <- droplevels(data$Clade, exclude= "Vombatiform")


#deleting all entries that have NAs in their main data and locomotor use; this means some species don't have weight data. T

data=data[-which(is.na( data$Tl_mm)) ,]
data=data[-which(is.na(data$Locomotor_use)),]

#make sure the species names are in the reows

rownames(data)<-data$Species


```


#import and process tree; sSynchronize trees and species by reducing tree to the data available; this has been checked for typos taht cause non-matches (fixed above in the code containing grep). The below culls species that are not in the tree from the data, and species that are not in the data from the tree.
```{r}
#134 species left; 14 newly identified species not present on tree but each of them are represented as close sister groups

tree=read.nexus("../Data/MayCollado_Marsupial_tree.nex")
namecheck=name.check(tree,data)
tree_synch=drop.tip(tree,namecheck$tree_not_data)
data_synch= data[-which(rownames(data) %in% namecheck$data_not_tree),]
name.check(tree_synch,data_synch)

#matching the order of tip labels to the tree

match(tree_synch$tip.label,data_synch$Species)

data_synch <-data_synch [,][match(tree_synch$tip.label,data_synch$Species),]

#Check that the re-ordering works

tree_synch$tip.label==data_synch$Species

#Some Petrogale branches are unresolved - hence, need to split these branches using multi2di and given a small branch length to allow for future use

is.binary(tree_synch)

tree_synch <- multi2di(tree_synch, random = FALSE)

tree_synch$edge.length[c( which (tree_synch$edge.length==0))] <- mean(tree_synch$edge.length)*0.1

#testing that the value is entered
tree_synch$edge.length[105]

is.binary(tree_synch)

```


#Processing data for partitions with m/f data available (assuming if m present, f is also present; but some species don't have male/female averages so they get deleted). 
```{r}
mf_data=data_synch[which(data_synch$Tl_Av_F !="NA"),]
mf_data=mf_data[which(mf_data$Wt_Av_F !="NA"),]


mf_datatree=drop.tip(tree_synch, name.check(tree_synch,mf_data)$tree_not_data )



```


#Packaging all outputs for analysis

```{r}

#putting all outputs and the initial processed input (dataFull) into one RDA 
dataFull<-data
data<-data_synch
tree<-tree_synch

save(dataFull,data,tree, mf_data,mf_datatree, file = "../Data/Processed_Data_for_Pgls.rda")

#clean up environment
remove(data_synch,tree_synch)

```
